Data: Use sufficient digits for new sort index
Let's say we're inserting a story between two stories
that have sort indices of 1.5 and 1.8. A simple / 2.0
calculation will give us a new index of 1.65.

However, we don't need to be that precise -- 1.6 will
suffice. This will allow us to more easily manage things
as the numbers get too tight and we run out of precision in
our data type.