Have server create ids when adding stories
When adding a new story into the backlog,
make the server be the only one who cares
about what the next story's id will be.

Also use CouchDB uuids for our ids instead
of ints. This is probably how we want our
data to be in the longer term, though we
might add a person-friendly id in the
future.