Multiple clients can create stories
Handle requests to create stories from
multiple clients by moving the first-story
calculation to the server side.

Works for situations where people are just
adding stories to the top of the backlog.

Does not work for situations where a story
is added on client A, moved, and a new
story is added on client B.