Fix: Backlog: Select story, not wrapper
When we started showing the story owner in
the top-level view, we made the story row
take up the width of the screen.

This lead to unexpected showing of the story
details, as clicks on supposed whitespace
were really hitting stories.

Anyway now only show the story if a click is
on the story element, and not the wrapper.