Backlog: Make draggable area same as visible area
The story wrappers were taking up the width of the
entire screen, while the visible width of the stories
is much less than that.

This was ok for desktops, but quickly became an issue
for touch devices.

Now the draggable area is the part of the story that
can be seen, and the whitespace on the sides is
actually whitespace.