Version 0.4
Ready for use by a small team.

Initialization steps are taken care of
automatically by the server on first run.

Cleaned up the UI a little bit, especially
for new-story entry.

Allows for a demo mode -- creating a user
with a publicly-visible password. Admittedly,
this is mostly only of use to the developer.

This release is probably fine for small-scale
deployment (i.e. one team), with only minor
rough edges (bugs) anticipated.