Fix: Hide comments on deadlines, keep colors
The gray in the last commit was too pervasive,
and included the background color of deadlines.

Hide comments for deadlines and the next-meeting
story. We might want to add these in some day,
but really if there is a question about a deadline
it should probably be addressed as an agenda item
or in some other way. Or, the time is just not right.