Backlog: Make comments go from bottom to top
Let's put the most-recent comment at the top.

In a discussion, it's useful to see what was
the last thing that was said. This also gives
us a solution to the closing-a-tall-story problem.

This might make the case for a document-link
field, though, that always stays near the top
of the story.