Fix: Backlog: Better data binding on story move
We were clearing out the stories list in the view
model and making a new array structure to data
bind to.

This is a lot of unnecessary work, and causes
the page to scroll to the top anyway whenever
we move, which is not what we want.

So, instead just update each element in the view
array with the correct element in order.