Add page to fix simple integrity issues
We can detect if there are simple issues
with the story linked list, like if there
are next-id loops, and we can fix them
with a button push, now.

Suggested fixes should still be reviewed
before they're put in place, and really
we should fix the code so this page is
no longer necessary, but it is here for now.