Version 0.7: Beta release
This commit marks the start of the 'beta era'
of Circle Blvd. development. It is useful enough
to be installed by technical audiences, used on
a daily basis to help small groups communicate,
and should work just fine for years, though it
might need to restart from time to time.

Be organized! :-)