Use 'compact' to minify JavaScript
For the time being, use the 'compact' module to
minify our JavaScript. Basically, this copies
our JavaScript files to public/_js, in minified
form, the first time they are accessed.

While this is a little useful, this is primarily 
to make Google happier with our site, so don't
try too hard to keep this system in place.