Introduce mindset: Roadmap
The 'roadmap' mindset is for easily putting
stories between two mileposts.

This first implementation puts together a
lot of our existing codebase -- basically
we just show another story list, that shares
the same data, but only shows mileposts.

The YUI drag and drop library happens to
"just work" for us in this scenario.