Version 0.8: Beta release 2
The second beta release focuses on the situations
when we want to add a lot of stories at once, or
move a lot of stories around.

It adds support for hashtags and filtering. It
introduces invitation links for adding people to
circles.

It introduces checklists, or saved lists of
stories, that can be added repeatedly.

It introduces different designs of the list, which
we're calling "mindsets," for those important
occasions when we want to organize everything, or
just decide what's next.

Thanks!