Drag and drop: Work with roadmap mindset
With the move to jQuery UI for drag and
drop, we had to make some adjustments for
the roadmap mindset to work out.

Primarily, we made a list of mileposts
to bind to in the view, which is probably
what we should have done in the first
place.

The rest was using the 'connectWith'
functionality of the Sortable.