Change 'move to top' to 'move above next meeting'
In practice, a 'move to top' button for deciding
what to work on next is annoying, since you
can't just think of what order you want to do
things in -- you have to continuously reorder
things after pressing the bump chevron.

So, let's try moving just above the Next Meeting
to see if it is a more natural behavior.