Version 1.0.1
This is closer to what 1.0 should have been. It
has some breaking changes.

If you're upgrading to this release, you'll need
to update your upstart script (now in deploy/amazon-ec2/)
as well as your git-deploy script, because we
changed the location of the files.

Also this release fixes the unit tests. The program
was working fine but the tests were broken.