Version 1.1.0
Introduce highlighted tasks that can be changed
with the arrow keys, working in a similar fashion
to a spreadsheet.

Tasks can also be moved via the keyboard, by way
of the standard cut and paste commands (for
example, Ctrl-X and Ctrl-V on Windows).

Press enter to open a highlighted task, and
escape to close it.