Fix: List: Drag / highlight interactions
There were a few things going on, here.

First, we weren't putting the DOM elements
in the right order after a drop event. This
is addressed with orderBlockInDom.

Then, postMoveBefore was sometimes wrong (if
the drag item / element wasn't at the start
of the block).

And, sometimes the highlighted block we got
at the end of a drag was different than the
one we started with, which caused a lot of
undesired behavior.