Story: Add icon for showing details on click
Since we moved to double-click to open stories,
there have been requests to do so with one click.

This approach still favors the "organizer" role
who is more interested in selecting tasks than
in opening them, but at the moment we're indeed
favoring organizers over people concerned solely
with their own tasks, though this may change in
the future.