Allow people who are not in circles to create them
If a person is removed from all circles, then they
should be able to sign in and create a circle of
their own.

This commit adds a page at /create which accesses
the circle creation API, and people are redirected
to it upon signin if they are not a member of any
circles.