Prelude: Add tips page
Add a page for the keyboard shortcuts. Also
move the 60-second demo to this page, and
call it the whirlwind tour.

Move the "slow" video back to the sign-in
page, as we expect it to be viewed more by
people who aren't really sure why they're
looking at the page.