Mainframe: Show circles' last archive date
As a site administrator, it's good to know
who is actively using the site. As archiving
tasks is a good indication of how active a
project is, let's use them to show the most
recent archive as well as the archive count.

This also means we don't need to make a
general activity stream right now.