Version 1.4.8
This is a usability and business metrics release.
Invitations are a little easier to understand, and
the tour has been updated to include more team
information.

We also added a 'welcome tour' for new accounts,
to introduce the task list explicitly.

And, there is now a report on the mainframe page
that shows how many members there are, how many
circles have been created, and what circles have
been active in the last 8 months.