Header: Remove title and menu
Start breaking down the header section. Remove
the title and menu, to decide where they want to
be. This means we can't sign out for now, so let's
keep this experiment out of master until further
work is accomplished.

Also highlight the active circle in blue, to see
how that goes.