Header: Add blue color, icons. Redesign.
Redesign the header based on customer
feedback. People want more color, so
we're giving it the 'info blue' background.

People keep looking at the top navbar
to sign out, so that's moved there, along
with the profile link.

The "What's next" mindset icon has been
removed (for now?), as it's typically
confusing and the roadmap has the function
covered.

More work needs to be done (in the coming
commits) on getting the rest of the
site to fit with this design, which may not
be final, but is in the direction we are
heading.