Header: Add icon for 'bump' mindset
This is only visible on screens larger
than sm and xs. I think this should go
into some alternative menu that we
haven't invented yet. But here it is today.

Also, make the size of the circle title
a little smaller, so we can fit the names
of some of our key customers on one line.