Refactor: Remove server-side archives experiment
A while ago we started to refactor the client-side
views and move them closer to the server side.

At this point it's more confusing than anything to keep
it around. The template is pretty simple, and most
of the work is no longer needed after porting the EJS
template to Jade.