Fix: Make design-doc creation idempotent
Extract design-doc saving into its own
module, and also don't mess with the design
docs array itself when creating them.

Failure to observe this best practice was
resulting in intermittent errors when
running tests, where we need to create
design documents with the same name in
different databases.

Our code was saying that documents had
been created when in fact they had not.