Rename Circle Blvd to Circle Tasks
Boulevard is hard to say, and the metaphor
was not helpful for anyone. People tended
to eventually call it "Circle."

Also, the street metaphor is overly restrictive,
and this rename will allow us to overhaul the
look more easily.

Circle Tasks also makes the "What does it do?"
conversation more straight forward.

Lastly, the Circle Blvd brand was not worth much,
making this change easy.