Hide 'Insert Story' when another task is highlighted
This UX is a compromise between what is easy to
do technically, and what we would like. Ideally
the 'Insert Story' would stay in place (and open)
as other tasks are selected and moved around;
however, since it is attached to the highlighted
story (for the time being), we hide the 'Insert
Story' panel if actions like story-move are
about to take place.